Switched to own mdns-cast-browser for discovery. Highly unstable!
