+++ castWebApi.js
@@ -1,8 +1,5 @@
 const http = require('http');
-const Client = require('castv2').Client;
-const Castv2Client = require('castv2-client').Client;
-const DefaultMediaReceiver = require('castv2-client').DefaultMediaReceiver;
-const mdns = require('mdns-js');
+const CastBrowser = require('../mdns-cast-browser/cast-browser');
 const url = require('url');
 const debug = require('debug')('cast-web-api');
 const args = require('minimist')(process.argv.slice(2));
@@ -10,14 +7,12 @@ const fetch = require('node-fetch');
 const os = require('os');
 const pkg = require('./package.json');
 const util = require('util');
-const events = require('events');
 const querystring = require('querystring');
 const chalk = require('chalk');
-const googleTTS = require('google-tts-api');
+const CastDevice = require('./cast-device');
 
 var hostname = '127.0.0.1';
 var port = 3000;
-var currentRequestId = 1;
 var timeoutDiscovery = 5000;
 var reconnectInterval = 300000;
 var discoveryInterval = 60000;
@@ -37,18 +32,17 @@ if (!windows) {
 
 function startApi() {
 	console.log('cast-web-api v'+thisVersion);
-	console.log('Discovering devices, please wait...');
-	discoverZones()
-	.then(devices => {
-		console.log('... done!');
-		setInterval(function() {
-			discoverZones();
-		}, discoveryInterval);
-		createWebServer();
-	})
-	.catch(errorMessage => {
-		console.log('Cannot discover devices, shutting down. Error message: '+ errorMessage);
-	})
+	
+	var browser = new CastBrowser();
+	browser.discover();
+
+	browser.on('deviceUp', device => {
+		console.log('deviceUp');
+		var newDevice = new CastDevice(device.id, device.address, device.name);
+		devices.push(newDevice);
+	});
+
+	createWebServer();
 }
 
 //HANDLE ARGUMENTS
@@ -452,586 +446,6 @@ function removeDevice(id) {
 	}
 }
 
-function CastDevice(id, address, name) {
-	this.id = id;
-	this.address = address;
-	this.name = name;
-	this.event = new events.EventEmitter();
-	this.link = 'disconnected';
-	this.reconnectInterval;
-	this.castConnectionReceiver;
-	this.castConnectionMedia;
-	this.status = {
-		volume: 0,
-		muted: false,
-		application: '',
-		status: '',
-		title: '',
-		subtitle: '',
-		image: ''
-	}
-
-	this.connect = function() {
-		connectReceiverCastDevice(this);
-		if (groupManagement) {
-			connectGroupMembers(this);
-		}
-	};
-
-    this.disconnect = function() {
-		disconnectReceiverCastDevice(this);
-	};
-
-	this.connectMedia = function() {
-		if (this.castConnectionReceiver.sessionId) {
-			connectMediaCastDevice(this, this.castConnectionReceiver.sessionId);
-		}
-	};
-
-	this.disconnectMedia = function() {
-		if (this.castConnectionMedia) {
-			disconnectMediaCastDevice(this);
-		}
-	};
-
-	this.toString = function() {
-		return {
-			id: this.id,
-			name: this.name,
-			connection: this.link,
-			address: this.address,
-			status: this.status,
-			groups: this.groups,
-			members: this.members
-		}
-	}
-
-	this.volume = function(targetLevel) {
-		log('info', 'CastDevice.volume()', targetLevel, this.id);
-		if (this.castConnectionReceiver.receiver && this.link == 'connected') {
-			this.castConnectionReceiver.receiver.send({ type: 'SET_VOLUME', volume: { level: targetLevel }, requestId: getNewRequestId() });
-			return {response:'ok'};
-		} else {
-			return {response:'error', error:'disconnected'};
-		}
-	}
-
-	this.muted = function(isMuted) {
-		log('info', 'CastDevice.muted()', isMuted, this.id);
-		if (this.castConnectionReceiver.receiver && this.link == 'connected') {
-			this.castConnectionReceiver.receiver.send({ type: 'SET_VOLUME', volume: { muted: isMuted }, requestId: getNewRequestId() });
-			return {response:'ok'};
-		} else {
-			return {response:'error', error:'disconnected'};
-		}
-	}
-
-	this.play = function() {
-		log('info', 'CastDevice.play()', '', this.id);
-		if (this.castConnectionMedia && this.link == 'connected') {
-			if (this.castConnectionMedia.media && this.castConnectionReceiver.sessionId && this.castConnectionMedia.mediaSessionId) {
-				this.castConnectionMedia.media.send({ type: 'PLAY', requestId: getNewRequestId(), mediaSessionId: this.castConnectionMedia.mediaSessionId, sessionId: this.castConnectionReceiver.sessionId });
-				return {response:'ok'};
-			} else {
-				return {response:'error', error:'nothing playing'};
-			}
-		} else {
-			return {response:'error', error:'nothing playing'};
-		}
-	}
-
-	this.pause = function() {
-		log('info', 'CastDevice.pause()', '', this.id);
-		if (this.castConnectionMedia && this.link == 'connected') {
-			if (this.castConnectionMedia.media && this.castConnectionReceiver.sessionId && this.castConnectionMedia.mediaSessionId) {
-				this.castConnectionMedia.media.send({ type: 'PAUSE', requestId: getNewRequestId(), mediaSessionId: this.castConnectionMedia.mediaSessionId, sessionId: this.castConnectionReceiver.sessionId });
-				return {response:'ok'};
-			} else {
-				return {response:'error', error:'nothing playing'};
-			}
-		} else {
-			return {response:'error', error:'nothing playing'};
-		}
-	}
-
-	this.stop = function() {
-		log('info', 'CastDevice.stop()', '', this.id);
-		if (this.castConnectionReceiver.sessionId && this.link == 'connected') {
-			this.castConnectionReceiver.receiver.send({ type: 'STOP', sessionId: this.castConnectionReceiver.sessionId, requestId: getNewRequestId() });
-			return {response:'ok'};
-		} else {
-			return {response:'error', error:'nothing playing'};
-		}
-	}
-
-	this.createSubscription = function(callback) {
-		createSubscription(this, callback);
-		return {response:'ok'};
-	}
-
-	this.removeSubscription = function() {
-		log('info', 'CastDevice.removeSubscription()', '', this.id);
-		//this.event.removeAllListeners('statusChange'); //Breaks group man
-		//this.event.removeAllListeners('linkChanged'); //Breaks reconnect man
-		delete this.callback;
-		return {response:'ok'};
-	}
-
-	this.setStatus = function(key, value) {
-		if (key=='volume' || key=='muted' || key=='application' || key=='status' || key=='title' || key=='subtitle' || key=='image' || key=='groupPlayback') {
-			if (value == null) { value = '' };
-			if (this.status[key] != value) {
-				this.status[key] = value;
-				this.event.emit('statusChange');
-			}
-		}
-	}
-
-	this.playMedia = function(media) {
-		playMediaCastDevice(this, media);
-		return {response:'ok'};
-	}
-
-	this.setZones = function(zones) {
-		setZonesCastDevice(this, zones);
-	}
-
-	reconnectionManagementInit(this);
-	subscriptionInit(this);
-	connectGroupMembersInit(this);
-}
-
-function connectReceiverCastDevice(castDevice) {
-	try {
-		log('info', 'CastDevice.connect()', 'host: ' + castDevice.address.host + ', port: ' + castDevice.address.port, castDevice.id );
-		castDevice.link = 'connecting';
-		castDevice.castConnectionReceiver = new Object();
-		castDevice.castConnectionReceiver.client = new Client();
-
-		castDevice.castConnectionReceiver.client.connect(castDevice.address, function() {
-			try {
-				castDevice.castConnectionReceiver.connection = castDevice.castConnectionReceiver.client.createChannel('sender-0', 'receiver-0', 'urn:x-cast:com.google.cast.tp.connection', 'JSON');
-			    castDevice.castConnectionReceiver.heartbeat = castDevice.castConnectionReceiver.client.createChannel('sender-0', 'receiver-0', 'urn:x-cast:com.google.cast.tp.heartbeat', 'JSON');
-			    castDevice.castConnectionReceiver.receiver = castDevice.castConnectionReceiver.client.createChannel('sender-0', 'receiver-0', 'urn:x-cast:com.google.cast.receiver', 'JSON');
-
-				castDevice.castConnectionReceiver.connection.send({ type: 'CONNECT' });
-				castDevice.castConnectionReceiver.receiver.send({ type: 'GET_STATUS', requestId: getNewRequestId() });
-
-			    castDevice.castConnectionReceiver.receiver.on('message', function(data, broadcast) {
-			    	parseReceiverStatusCastDevice(castDevice, data);
-			    	if (castDevice.link != 'connected') {
-						castDevice.link = 'connected';
-						castDevice.event.emit('linkChanged');
-			    	}
-			   	});
-
-			   	castDevice.castConnectionReceiver.receiver.on('error', function(error) {
-			   		log('error', 'CastDevice.connect()', 'castDevice.castConnectionReceiver.receiver error: '+error, castDevice.id);
-					castDevice.disconnect();
-			   	});
-
-			   	castDevice.castConnectionReceiver.connection.on('error', function(error) {
-			   		log('error', 'CastDevice.connect()', 'castDevice.castConnectionReceiver.connection error: '+error, castDevice.id);
-					castDevice.disconnect();
-			   	});
-
-			   	castDevice.castConnectionReceiver.heartBeatIntervall = setInterval(function() {
-					if (castDevice.castConnectionReceiver) {
-						try {
-							castDevice.castConnectionReceiver.heartbeat.send({ type: 'PING' });
-						} catch (e) {
-							log('error', 'CastDevice.connect() castConnectionReceiver.heartBeatIntervall', 'exception: '+e, castDevice.id);
-							castDevice.disconnect(); //TODO:
-						}
-						
-					}
-				}, 5000);
-			} catch (e) {
-				log('error', 'CastDevice.connect()', 'exception: '+e, castDevice.id);
-				castDevice.disconnect();
-			}
-		});
-		castDevice.castConnectionReceiver.client.on('error', function(err) {
-		 	log('error', 'CastDevice.connect()', 'castDevice.castConnectionReceiver.client error: '+err, castDevice.id);
-		 	castDevice.disconnect();
-		});
-	} catch (e) {
-		log('error', 'CastDevice.connect()', 'exception: '+e, castDevice.id);
-		castDevice.disconnect();
-	}
-}
-
-function disconnectReceiverCastDevice(castDevice) {
-	log('info', 'CastDevice.disconnect()', 'host: ' + castDevice.address.host + ', port: ' + castDevice.address.port, castDevice.id );
-	try {
-		if (castDevice.link != 'disconnected') {
-			castDevice.link = 'disconnected';
-			castDevice.event.emit('linkChanged');
-			clearInterval(castDevice.castConnectionReceiver.heartBeatIntervall);
-			//castDevice.castConnectionReceiver.connection.send({ type: 'CLOSE' });
-			castDevice.castConnectionReceiver.client.close();
-			//castDevice.castConnectionReceiver = null;
-		}
-	} catch (e) {
-		//castDevice.castConnectionReceiver = null;
-		log('error', 'CastDevice.disconnect()', 'exception: '+e, castDevice.id);
-	}
-	castDevice.disconnectMedia();
-}
-
-function reconnectionManagementInit(castDevice) {
-	log('debug', 'reconnectionManagementInit()', '', castDevice.id);
-
-	castDevice.event.on('linkChanged', function() {
-		reconnectionManagement(castDevice);
-	});
-}
-
-function reconnectionManagement(castDevice) {
-	log('debug', 'reconnectionManagement()', 'link changed to: ' + castDevice.link + ', reconnectInterval: ' + castDevice.reconnectInterval, castDevice.id);
-	if (castDevice.link=='disconnected') {
-		if (castDevice.reconnectInterval==null) {
-			log('debug', 'reconnectionManagement()', 'starting interval', castDevice.id);
-			castDevice.reconnectInterval = setInterval(function() {
-				log('debug', 'reconnectionManagement()', 'reconnect evaluating', castDevice.id);
-				if (castDevice.link!='connected') {
-					log('info', 'reconnectionManagement()', 'reconnecting', castDevice.id);
-					castDevice.connect();
-				}
-			}, reconnectInterval);
-		}
-	} else {
-		log('debug', 'reconnectionManagement()', 'interval evaluating', castDevice.id);
-		if (castDevice.reconnectInterval!=null) {
-			log('debug', 'reconnectionManagement()', 'interval remove', castDevice.id);
-			clearInterval(castDevice.reconnectInterval);
-			castDevice.reconnectInterval = null;
-		}
-	}
-}
-
-function connectMediaCastDevice(castDevice, sessionId) {
-	try {
-		log('info', 'CastDevice.connectMedia()', 'host: ' + castDevice.address.host + ', port: ' + castDevice.address.port + ', sessionId: ' + sessionId, castDevice.id);
-		castDevice.castConnectionMedia = new Object();
-		castDevice.castConnectionMedia.client = new Client();
-
-		castDevice.castConnectionMedia.client.connect(castDevice.address, function() {
-			try {
-				castDevice.castConnectionMedia.connection = castDevice.castConnectionMedia.client.createChannel('sender-0', sessionId, 'urn:x-cast:com.google.cast.tp.connection', 'JSON');
-				castDevice.castConnectionMedia.heartbeat = castDevice.castConnectionMedia.client.createChannel('sender-0', sessionId, 'urn:x-cast:com.google.cast.tp.heartbeat', 'JSON');
-				castDevice.castConnectionMedia.media = castDevice.castConnectionMedia.client.createChannel('sender-0', sessionId, 'urn:x-cast:com.google.cast.media', 'JSON');
-
-				castDevice.castConnectionMedia.connection.send({ type: 'CONNECT' });
-				castDevice.castConnectionMedia.media.send({ type: 'GET_STATUS', requestId: getNewRequestId() });
-
-				castDevice.castConnectionMedia.media.on('message', function(data, broadcast) {
-					parseMediaStatusCastDevice(castDevice, data);
-					castDevice.castConnectionMedia.link = 'connected';
-				});
-
-				castDevice.castConnectionMedia.media.on('error', function(error) {
-					log('error', 'CastDevice.connectMedia()', 'castDevice.castConnectionMedia.media error: '+error, castDevice.id);
-					castDevice.disconnectMedia();
-				});
-
-				castDevice.castConnectionMedia.connection.on('error', function(error) {
-					log('error', 'CastDevice.connectMedia()', 'castDevice.castConnectionMedia.connection error: '+error, castDevice.id);
-					castDevice.disconnectMedia();
-				})
-
-				castDevice.castConnectionMedia.heartBeatIntervall = setInterval(function() {
-					try {
-						if (castDevice.castConnectionMedia && castDevice.link=='connected') {
-							castDevice.castConnectionMedia.heartbeat.send({ type: 'PING' });
-						}
-					} catch(e) {
-						log('error', 'CastDevice.connectMedia()', 'heartbeat exception: '+e, castDevice.id);
-						castDevice.disconnectMedia();
-					}
-				}, 5000);
-			} catch(e) {
-				log('error', 'CastDevice.connectMedia()', 'exception: '+e, castDevice.id);
-				castDevice.disconnectMedia();
-			}
-		});
-		castDevice.castConnectionMedia.client.on('error', function(err) {
-			log('error','CastDevice.connectMedia()', 'castDevice.castConnectionMedia.client error: '+err, castDevice.id);
-			castDevice.disconnectMedia();
-		});
-	} catch(e) {
-		log('error', 'CastDevice.connectMedia()', 'exception: '+e, castDevice.id);
-		castDevice.disconnectMedia();
-	}
-}
-
-function disconnectMediaCastDevice(castDevice) {
-	log('info', 'CastDevice.disconnectMedia()', 'host: ' + castDevice.address.host + ', port: ' + castDevice.address.port, castDevice.id);
-	try {
-		if (castDevice.castConnectionMedia!=null) {
-			clearInterval(castDevice.castConnectionMedia.heartBeatIntervall);
-			//castDevice.castConnectionMedia.connection.send({ type: 'CLOSE' });
-			castDevice.castConnectionMedia.client.close();
-			castDevice.castConnectionMedia.link = 'disconnected';
-			castDevice.status.status = '';
-			castDevice.status.title = '';
-			castDevice.status.subtitle = '';
-			castDevice.status.image = '';
-			castDevice.event.emit('statusChange');
-			//castDevice.castConnectionMedia = null;
-		}
-	} catch(e) {
-		log('error', 'CastDevice.disconnectMedia()', 'exception: '+e, castDevice.id); //TODO: Notify subscriber
-		//castDevice.castConnectionMedia = null;
-	}
-}
-
-function parseReceiverStatusCastDevice(castDevice, receiverStatus) {
-	try {
-		log('debug', 'parseReceiverStatusCastDevice()', 'receiverStatus: ' + JSON.stringify(receiverStatus), castDevice.id );
-
-		if (receiverStatus.type == 'RECEIVER_STATUS') {
-			if (receiverStatus.status.applications) {
-				if ( receiverStatus.status.applications[0].sessionId != castDevice.castConnectionReceiver.sessionId) {
-					log('debug', 'parseReceiverStatusCastDevice()', 'sessionId changed', castDevice.id);
-					if (receiverStatus.status.applications[0].isIdleScreen!=true) {
-						log('debug', 'parseReceiverStatusCastDevice()', 'isIdleScreen: '+receiverStatus.status.applications[0].isIdleScreen+', sessionId changed to: '+receiverStatus.status.applications[0].sessionId+', from: '+castDevice.castConnectionReceiver.sessionId, castDevice.id);
-						castDevice.castConnectionReceiver.sessionId = receiverStatus.status.applications[0].sessionId;
-						castDevice.connectMedia();
-					}
-				}
-				if ( receiverStatus.status.applications[0].displayName ) {
-					castDevice.setStatus('application', receiverStatus.status.applications[0].displayName);
-				}
-			} else {
-				castDevice.castConnectionReceiver.sessionId = null;
-				castDevice.status.application = '';
-				castDevice.disconnectMedia();
-			}
-			if (receiverStatus.status.volume) {
-				castDevice.setStatus( 'volume', Math.round(receiverStatus.status.volume.level*100) );
-				castDevice.setStatus('muted', receiverStatus.status.volume.muted);
-			}
-		}
-	} catch(e) {
-		log('error', 'parseReceiverStatusCastDevice()', 'exception: '+e, castDevice.id);
-	}
-}
-
-function parseMediaStatusCastDevice(castDevice, mediaStatus) {
-	try {
-		log('debug', 'parseMediaStatusCastDevice()', 'mediaStatus: ' + JSON.stringify(mediaStatus), castDevice.id );
-
-		if (mediaStatus.type == 'MEDIA_STATUS') {
-			if (mediaStatus.status[0]) {
-				if (mediaStatus.status[0].media) {
-					if (mediaStatus.status[0].media.metadata) {
-						var metadataType = mediaStatus.status[0].media.metadata.metadataType;
-						if(metadataType<=1) {
-							castDevice.setStatus('title', mediaStatus.status[0].media.metadata.title);
-							castDevice.setStatus('subtitle', mediaStatus.status[0].media.metadata.subtitle);
-						} 
-						if(metadataType==2) {
-							castDevice.setStatus('title', mediaStatus.status[0].media.metadata.seriesTitle);
-							castDevice.setStatus('subtitle', mediaStatus.status[0].media.metadata.subtitle);
-						} 
-						if(metadataType>=3 && metadataType<=4) {
-							castDevice.setStatus('title', mediaStatus.status[0].media.metadata.title);
-							castDevice.setStatus('subtitle', mediaStatus.status[0].media.metadata.artist);
-						}
-						if (metadataType>=0 && metadataType<4) {
-							if (mediaStatus.status[0].media.metadata.images) {
-								if (mediaStatus.status[0].media.metadata.images[0]) {
-									castDevice.setStatus('image', mediaStatus.status[0].media.metadata.images[0].url);
-								}
-							}
-						}
-					}
-				}
-
-				if (mediaStatus.status[0].mediaSessionId) {
-					castDevice.castConnectionMedia.mediaSessionId = mediaStatus.status[0].mediaSessionId;
-				}
-
-				if (mediaStatus.status[0].playerState) {
-					castDevice.setStatus('status', mediaStatus.status[0].playerState);
-				}
-			}
-		}
-	} catch(e) {
-		log('error', 'parseMediaStatusCastDevice()', 'exception: '+e, castDevice.id);
-	}
-}
-
-function playMediaCastDevice(castDevice, media) {
-	try {
-		log('info', 'CastDevice.playMedia()', 'media: ' + JSON.stringify(media), castDevice.id);
-
-		var castv2Client = new Castv2Client();
-		var mediaList = [];
-
-		media.forEach(function(element, index) {
-			var mediaElement =  {
-				autoplay : true,
-				preloadTime : 5,
-				activeTrackIds : [],
-				googleTTS: element.googleTTS,
-				//playbackDuration: 4,
-				//startTime : 1,
-				media: {
-					contentId: element.mediaUrl,
-					contentType: element.mediaType,
-					streamType: element.mediaStreamType,
-					metadata: {
-						type: 0,
-						metadataType: 0,
-						title: element.mediaTitle,
-						subtitle: element.mediaSubtitle,
-						images: [ { url: element.mediaImageUrl } ]
-					}
-				}
-			};
-			log('info', 'CastDevice.playMedia()', 'mediaElement: ' + JSON.stringify(mediaElement), castDevice.id);
-			mediaList.push(mediaElement);
-		});
-
-		replaceGoogleTts(mediaList)
-		.then(newMediaList => {
-			log('info', 'CastDevice.playMedia()', 'newMediaList: ' + JSON.stringify(newMediaList), castDevice.id);
-		
-		  	castv2Client.connect(castDevice.address, function() {
-				castv2Client.launch(DefaultMediaReceiver, function(err, player) {
-					player.queueLoad(newMediaList, {startIndex:0, repeatMode: "REPEAT_OFF"}, function(err, status) {
-						log('info', 'CastDevice.playMedia()', 'loaded queue: ' + status, castDevice.id);
-					});
-			    });
-		 	});
-
-		 	setTimeout(() => {
-				try{ castv2Client.close(); } catch(e) { log('error', 'CastDevice.playMedia()', 'castv2Client.close() exception: '+e, castDevice.id ); }
-			}, 5000);
-		})
-		.catch(error => {
-			log('error', 'CastDevice.playMedia()', 'replaceGoogleTts error: '+error, castDevice.id );
-		})
-	} catch(e) {
-		log('error', 'CastDevice.playMedia()', 'exception: '+e, castDevice.id );
-	}
-}
-
-function replaceGoogleTts(mediaList) {
-	return new Promise( function(resolve, reject) {
-		var googleTTSPromised = 0;
-		var googleTTSResolved = 0;
-
-		mediaList.forEach(function(mediaElement){
-			if (mediaElement.googleTTS && mediaElement.googleTTS!=null) {
-				googleTTSPromised++;
-			} else {
-				delete mediaElement.googleTTS
-			}
-		});
-
-		mediaList.forEach(function(mediaElement){
-			if (mediaElement.googleTTS && mediaElement.googleTTS!=null) {
-				googleTTS(mediaElement.media.metadata.title, mediaElement.googleTTS, 1)
-				.then(function (url) {
-					mediaElement.media.contentId = url;
-					mediaElement.media.mediaType = 'audio/mp3';
-					mediaElement.media.mediaStreamType = 'BUFFERED';
-					googleTTSResolved++;
-					if (googleTTSResolved == googleTTSPromised) {
-						resolve(mediaList);
-					}
-				})
-				.catch(function (err) {
-					log('error', 'replaceGoogleTts()', 'googleTTS error: '+err);
-					delete mediaElement;
-					googleTTSPromised--;
-					if (googleTTSResolved == googleTTSPromised) {
-						resolve(mediaList);
-					}
-				})
-			}
-		});
-
-		if (googleTTSResolved == googleTTSPromised) {
-			resolve(mediaList);
-		}
-
-		setTimeout(function() {
-			reject('Google TTS timeout.');
-		}, 5000);
-	});
-}
-
-function subscriptionInit(castDevice) {
-	log('debug', 'CastDevice.subscriptionInit()', '', castDevice.id);
-	castDevice.event.on('statusChange', function() {
-		if (castDevice.callback) {
-			//log('info', 'CastDevice.subscriptionInit()', 'castDevice.toString(): '+JSON.stringify( castDevice.toString() )+', '+JSON.stringify( castDevice.callback ), castDevice.id);
-			sendCallBack( castDevice.toString(), castDevice.callback );
-		}
-	});
-
-	castDevice.event.on('linkChanged', function() {
-		if (castDevice.callback) {
-			//log('info', 'CastDevice.subscriptionInit()', 'castDevice.toString(): '+JSON.stringify( castDevice.toString() )+', '+JSON.stringify( castDevice.callback ), castDevice.id);
-			sendCallBack( castDevice.toString(), castDevice.callback );
-		}
-	});
-}
-
-function createSubscription(castDevice, callback) {
-	castDevice.removeSubscription();
-
-	try {
-		castDevice.callback = url.parse('http://'+callback);
-	} catch(e) {
-		delete castDevice.callback;
-	}
-	
-	log('info', 'CastDevice.createSubscription()', 'callback: '+ JSON.stringify(castDevice.callback), castDevice.id);
-	if (castDevice.callback) {
-		sendCallBack( castDevice.toString(), castDevice.callback );
-	}
-}
-
-function sendCallBack(status, callback) {
-	log( 'debug', 'sendCallBack()', 'to: '+ JSON.stringify(callback) +', status: ' + JSON.stringify(status), status.id );
-
-	try{
-		var data = JSON.stringify(status);
-
-		var options = {
-			hostname: callback.hostname,
-			port: callback.port,
-			path: callback.path,
-			method: 'POST',
-			headers: {
-				'Content-Type': 'application/json',
-				'Content-Length': Buffer.byteLength(data)
-			}
-		};
-
-		var req = http.request(options, function(res) {
-			res.setEncoding('utf8');
-			/*res.on('data', function (chunk) {
-				console.log("Answer: " + chunk);
-			});*/
-		});
-
-		req.on('error', function(error) {
-			log('error', 'sendCallBack()', 'cannot send callback: ' + JSON.stringify(callback) + ', error: ' + error, status.id);
-		});
-
-		req.write(data);
-		req.end();
-	} catch (e) {
-		log('error', 'sendCallBack()', 'cannot send callback: ' + JSON.stringify(callback) + ', error: ' + error, status.id);
-	}
-}
-
 function getRestOfPathArray(pathArray, start) {
 	var restOfPathArray = '';
 	pathArray.forEach(function(element, index) {
@@ -1047,490 +461,6 @@ function getRestOfPathArray(pathArray, start) {
 	} else {
 		return restOfPathArray;
 	}
-}
-
-function discoverZones() {
-	return new Promise( function(resolve, reject) {
-		if ( getNetworkIp() ) {
-			discoverTimes(discoveryRuns, [])
-			.then(zoneResults => {
-				var merged = [];
-				zoneResults.forEach(function(zoneResult) {
-					zoneResult.zones.forEach(function(zone) {
-						var exists = false;
-
-						merged.forEach(function(mergedZone) {
-							if (mergedZone.id == zone.id) {
-								exists = true;
-								if (zone.groups) {
-									zone.groups.forEach(function(groupId) {
-										if (mergedZone.groups.indexOf(groupId) < 0) {
-											mergedZone.groups.push(groupId);
-										}
-									});
-								}
-							}
-						});
-						
-						if (!exists) {
-							merged.push(zone);
-							log('debug', 'discoverZones()', 'doesnt exist, pushing: ' + zone.id);
-						} else {
-							log('debug', 'discoverZones()', 'exists, pushed already: ' + zone.id);
-						}
-					});
-				});
-				createGoogleZones(merged);
-				resolve(merged);			
-			})
-			.catch(errorMessage => {
-				log('error', 'discoverZones()', 'failed: ' + errorMessage);
-				reject(errorMessage);
-			})
-		} else {
-			resolve({});
-		}
-	});
-}
-
-function discoverTimes(times, zoneResults) {
-	return new Promise( function(resolve, reject) {
-		discover('googlezone')
-		.then(devices => {
-			log('debug', 'discoverTimes()', 'discover("googlezone") done times: '+times+', zoneResults: '+zoneResults+', typeof: '+ typeof zoneResults);
-			zoneResults.push({ zones: devices });
-			if (times>1) {
-				log( 'debug', "discoverTimes()", 'newZoneResults: '+zoneResults+', new times: '+(times-1) );
-				resolve( discoverTimes( (times-1), zoneResults ) );
-			} else {
-				discover('googlecast')
-				.then(devices => {
-					resolve( zoneResults );
-				})
-				.catch(errorMessage => {
-					log('error', 'discoverTimes()', 'discover("googlecast") failed: ' + errorMessage);
-					resolve( zoneResults );
-				})
-			}
-		})
-		.catch(errorMessage => {
-			log('error', 'discoverTimes()', 'discover("googlezone") failed: ' + errorMessage);
-			reject(errorMessage);
-		})
-	});
-}
-
-function existsInMerged(merged, id) {
-	var exists = false;
-	merged.forEach(function(zone) {
-		if (zone.id == id) {
-			exists = true;
-		}
-	});
-	return exists;
-}
-
-//GOOGLE CAST FUNCTIONS 'googlecast' 'googlezone'
-function discover(target) {
-	var both = false;
-	if (!target) {
-		target = 'googlecast';
-		both = true;
-	}
-	return new Promise( function(resolve, reject) {
-		var updateCounter=0;
-		var discovered = [];
-		try {
-			if (getNetworkIp) {
-				var browser = mdns.createBrowser(mdns.tcp(target));
-				var exception;
-
-				browser.on('error', function(error) {
-					log('debug', 'discover()', 'mdns browser error: ' + error);
-				})
-
-				browser.on('ready', function(){
-					browser.discover();
-				});
-
-				browser.on('update', function(service){
-					try {
-						updateCounter++;
-						log('debug', 'discover()', 'update received, service: ' + JSON.stringify(service));
-						if (target=='googlecast' && service.type[0].name==target) {
-							var currentDevice = {
-								id: getId(service.txt[0]),
-								name: getFriendlyName(service.txt),
-									address: {
-										host: service.addresses[0],
-										port: service.port
-								}
-							}
-					  		if (!duplicateDevice(discovered, currentDevice) && currentDevice.name!=null ) {
-					  			log('debug', 'discover()', 'found device: '+ JSON.stringify(currentDevice));
-					  			discovered.push(currentDevice);
-					  			updateExistingCastDeviceAddress(currentDevice);
-
-					  			if ( !deviceExists(currentDevice.id) ) {
-					  				log('info', 'discover()', 'added device name: '+ currentDevice.name +', address: '+ JSON.stringify(currentDevice.address), currentDevice.id);
-					  				devices.push( new CastDevice( currentDevice.id, currentDevice.address, currentDevice.name ) ); //TODO: addDevice
-					  			}
-					  		} else {
-					  			log('debug', 'discover()', 'duplicate, googlezone device or empy name: ' + JSON.stringify(currentDevice));
-					  		}
-						}
-						if (target=='googlezone' && service.type[0].name==target) {
-							var currentGroupMembership = {
-								id: getId(service.txt[0]).replace(/-/g, ''),
-								groups: getGroupIds(service.txt)
-							}
-							log('debug', 'discover()', 'found googlezone: ' + JSON.stringify(currentGroupMembership) );
-							discovered.push(currentGroupMembership);
-						}
-				  	} catch (e) {
-						log('error', 'discover()', 'exception while prcessing service: '+e);
-					}
-				});
-			}
-		} catch (e) {
-			reject('Exception caught: ' + e);
-		}
-
-		setTimeout(() => {
-			try{
-				browser.stop();
-			} catch (e) {
-				//reject('Exception caught: ' + e)
-			}
-			log('debug', 'discover()', 'updateCounter: ' + updateCounter);
-			resolve(discovered);
-	  	}, timeoutDiscovery);
-	});
-}
-
-function updateExistingCastDeviceAddress(discoveredDevice) {
-	log('debug', 'updateExistingCastDeviceAddress()', 'discoveredDevice: ' + JSON.stringify(discoveredDevice), discoveredDevice.id );
-	if ( deviceExists(discoveredDevice.id) ) {
-		var castDevice = getDevice(discoveredDevice.id);
-		log('debug', 'updateExistingCastDeviceAddress()', 'exists', discoveredDevice.id);
-		castDevice.name = discoveredDevice.name;
-		if (discoveredDevice.address.host && discoveredDevice.address.port) {
-			if (castDevice.address.host != discoveredDevice.address.host || castDevice.address.port != discoveredDevice.address.port) {
-				log('info', 'updateExistingCastDeviceAddress()', 'updating address from: '+castDevice.address.host+':'+castDevice.address.port+' to: '+discoveredDevice.address.host+':'+discoveredDevice.address.port, discoveredDevice.id);
-
-				if (castDevice.link == 'connected') {
-					castDevice.disconnect();
-					castDevice.event.once('linkChanged', function() {
-						log('info', 'updateExistingCastDeviceAddress()', 'once linkChanged: ' + castDevice.link, castDevice.id);
-						castDevice.address = {
-							host: discoveredDevice.address.host,
-							port: discoveredDevice.address.port
-						};
-					});
-				} else {
-					castDevice.address = {
-						host: discoveredDevice.address.host,
-						port: discoveredDevice.address.port
-					};
-				}
-			}
-		}
-	}
-}
-
-function createGoogleZones(discoveredZones) {
-	log( 'debug', 'createGoogleZones()', 'discoveredZones: ' + JSON.stringify(discoveredZones) );
-	var zones = discoveredZones;
-
-	discoveredZones.forEach(function(device) {
-		if (device.groups) {
-			device.groups.forEach(function(groupId) {
-				var groupExists = false;
-				zones.forEach(function(element) {
-					if (element.id == groupId) {
-						groupExists = element;
-					}
-				});
-				if (groupExists) {
-					log( 'debug', 'createGoogleZones()', 'groupExists: ' + JSON.stringify(groupExists) );
-					groupExists.members.push(device.id);
-				} else {
-					log( 'debug', 'createGoogleZones()', 'group doesnt exist: ' + groupExists );
-					zones.push({
-						id: groupId,
-						members: [ device.id ]
-					});
-				}
-			});
-		}
-	});
-
-	log( 'debug', 'createGoogleZones()', 'done! zones: ' + JSON.stringify(zones) );
-	devices.forEach(function(device) {
-		var zone = { id: device.id };
-		
-		zones.forEach(function(element) {
-			if (element.id == device.id) {
-				zone = element;
-			}
-		});
-
-		device.setZones(zone);
-	})
-}
-
-function setZonesCastDevice(castDevice, zones) {
-	log( 'debug', 'setZonesCastDevice()', 'zones: ' + JSON.stringify(zones), castDevice.id );
-	
-	if (zones.members) {
-		if (castDevice.members) {
-			zones.members.forEach(function(memberId) {
-				if ( castDevice.members.indexOf(memberId) < 0 ) {
-					//log( 'info', 'setZonesCastDevice()', 'added ('+ castDevice.name +') group member: ' + memberId, castDevice.id );
-					castDevice.event.emit('groupChange', 'add', { id: castDevice.id, members: [memberId] });
-				}
-			});
-			castDevice.members.forEach(function(memberId) {
-				if ( zones.members.indexOf(memberId) < 0 ) {
-					//log( 'error', 'setZonesCastDevice()', 'removed ('+ castDevice.name +') group member: ' + memberId, castDevice.id );
-					castDevice.event.emit('groupChange', 'remove', { id: castDevice.id, members: [memberId] });
-				}
-			});
-		} else {
-			//log( 'info', 'setZonesCastDevice()', 'added ('+ castDevice.name +') group members: ' + zones.members, castDevice.id );
-			castDevice.event.emit('groupChange', 'add', { id: castDevice.id, members: zones.members });
-		}
-		castDevice.members = zones.members;
-	} else {
-		if (castDevice.members) {
-			//log( 'error', 'setZonesCastDevice()', 'removed ('+ castDevice.name +') group members: ' + castDevice.members, castDevice.id );
-			castDevice.event.emit('groupChange', 'remove', { id: castDevice.id, members: castDevice.members });
-		}
-		delete castDevice.members;
-	}
-
-	if (zones.groups) {
-		if (castDevice.groups) {
-			zones.groups.forEach(function(groupId) {
-				if ( castDevice.groups.indexOf(groupId) < 0 ) {
-					//log( 'info', 'setZonesCastDevice()', 'added ('+ castDevice.name +') group: ' + groupId, castDevice.id );
-					castDevice.event.emit('groupChange', 'add', { id: castDevice.id, groups: [groupId] });
-				}
-			});
-			castDevice.groups.forEach(function(groupId) {
-				if ( zones.groups.indexOf(groupId) < 0 ) {
-					//log( 'info', 'setZonesCastDevice()', 'removed ('+ castDevice.name +') group: ' + groupId, castDevice.id );
-					castDevice.event.emit('groupChange', 'remove', { id: castDevice.id, groups: [groupId] });
-				}
-			});
-		} else {
-			//log( 'info', 'setZonesCastDevice()', 'added ('+ castDevice.name +') groups: ' + zones.groups, castDevice.id );
-			castDevice.event.emit('groupChange', 'add', { id: castDevice.id, groups: zones.groups });
-		}
-		castDevice.groups = zones.groups;
-	} else {
-		if (castDevice.groups) {
-			//log( 'error', 'setZonesCastDevice()', 'removed ('+ castDevice.name +') groups: ' + castDevice.groups, castDevice.id );
-			castDevice.event.emit('groupChange', 'remove', { id: castDevice.id, groups: castDevice.groups });
-		}
-		delete castDevice.groups;
-	}
-
-}
-
-function connectGroupMembersInit(castDevice) {
-	castDevice.event.on('statusChange', function() {
-		if (castDevice.members) {
-			if (castDevice.status.application) {
-				if (castDevice.status.application!='Backdrop' && castDevice.status.application!='') { //TODO: use isIdleScreen
-					log( 'info', 'connectGroupMembersInit()', 'statusChange is group', castDevice.id );
-					syncGroupMemberStatus(castDevice.status, castDevice.members, castDevice.id, true);
-				} else {
-					log( 'error', 'connectGroupMembersInit()', 'statusChange is APP no title group', castDevice.id );
-					resetGroupMemberStatus(castDevice);
-				}
-			} else {
-				log( 'error', 'connectGroupMembersInit()', 'statusChange is NO APP group', castDevice.id );
-				resetGroupMemberStatus(castDevice);
-			}
-		}
-	});
-
-	castDevice.event.on('groupChange', function(operation, zone) {
-		var level = 'info';
-
-		if (operation == 'remove') {
-			level = 'error';
-		}
-
-		if (zone.members) {
-			log( level, 'on groupChange()', operation+' members: ' + zone.members, zone.id );
-			if (operation == 'remove') {
-				zone.members.forEach(function(memberId) {
-					if ( deviceExists(memberId) ) {
-						var member = getDevice(memberId);
-						if (member.groupPlayback) {
-							log( level, 'on groupChange()', 'resetting groupPlayback for member: ' + memberId, zone.id );
-							syncGroupMemberStatus({ application: '', status: '', title: '', subtitle: '', image: '' }, [memberId], zone.id, false); //Why []?
-						}
-					}
-				});
-			}
-			if (operation == 'add') {
-				if ( deviceExists(zone.id) ) {
-					var group = getDevice(zone.id);
-					log( level, 'on groupChange()', 'evaluating groupPlayback for new members: ' + zone.members, zone.id );
-					group.event.emit('statusChange'); //maybe delay, coz event call before actual castDevice change
-				}
-			}
-		}
-
-		if (zone.groups) {
-			log( level, 'on groupChange()', operation+' groups: ' + zone.groups, zone.id );
-		}
-	});
-}
-
-function connectGroupMembers(castDevice) {
-	if (castDevice.members) {
-		castDevice.members.forEach(function(member) {
-			if ( deviceExists(member) && member && member!='' ) {
-				var castDeviceMember = getDevice(member);
-				if (castDeviceMember.link=='disconnected') {
-					log( 'info', 'connectGroupMembers()', 'connecting member: ' + castDeviceMember.id + ', castDeviceMember.link: '+castDeviceMember.link, castDevice.id );
-					//castDeviceMember.link = 'connecting';
-					castDeviceMember.connect();
-				}
-			}
-		});
-	}
-
-	if (castDevice.groups) {
-		castDevice.groups.forEach(function(group) {
-			if ( group && group!='' && deviceExists(group) ) {
-				var castDeviceGroup = getDevice(group);
-				if (castDeviceGroup.link=='disconnected') {
-					log( 'info', 'connectGroupMembers()', 'connecting group: ' + group + ', castDeviceGroup.link: '+castDeviceGroup.link, castDevice.id );
-					//castDeviceGroup.link = 'connecting';
-					castDeviceGroup.connect();
-				}
-			}
-		});
-	}
-}
-
-function resetGroupMemberStatus(castDevice) {
-	if (castDevice.members) { if (castDevice.members[0]) { if ( deviceExists(castDevice.members[0]) ) {
-		var member = getDevice(castDevice.members[0]);
-		if ( member.status ) {
-			if (member.status.groupPlaybackOn == castDevice.id) {
-				syncGroupMemberStatus({ application: '', status: '', title: '', subtitle: '', image: '' }, castDevice.members, castDevice.id, false);
-			} else {
-				log( 'info', 'resetGroupMemberStatus()', 'not resetting, getDevice(castDevice.members[0]).status.groupPlaybackOn: '+getDevice(castDevice.members[0]).status.groupPlaybackOn+' != '+castDevice.id, castDevice.id );
-			}
-		} else {
-			log( 'error', 'resetGroupMemberStatus()', 'device.status: '+device.status, castDevice.id );
-		}
-		
-	} } }
-	
-}
-
-function syncGroupMemberStatus(status, members, id, groupPlayback) {
-	log( 'info', 'syncGroupMemberStatus('+groupPlayback+')', 'to members: '+ members +', new status: '+ JSON.stringify(status), id);
-
-	members.forEach(function(member) {
-		if ( deviceExists(member) || member || member!='' ) {
-			var castDeviceMember = getDevice(member);
-			if (groupPlayback) {
-				castDeviceMember.status.groupPlayback = groupPlayback;
-				castDeviceMember.status.groupPlaybackOn = id;
-			} else {
-				delete castDeviceMember.status.groupPlayback;
-				delete castDeviceMember.status.groupPlaybackOn;
-			}
-			for (var key in status) {
-				if (key != 'volume' && key != 'muted') {
-					castDeviceMember.setStatus(key, status[key]);
-				}
-			}
-		}
-	});
-}
-
-function setGroupLevel(castDevice, level) {
-	log('info', 'setGroupLevel()', 'target level: '+ level, castDevice.id);
-	if (castDevice.members) {
-		castDevice.members.forEach(function(member) {
-			if ( deviceExists(member) || member || member!='' ) {
-				getDevice(member).volume(level);
-			}
-		});
-	}
-}
-
-function duplicateDevice(devices, device) {
-	if (device.id && device.id!=null && devices && devices!=null) {
-		for (var i = 0; i < devices.length; i++) {
-			if(devices[i].id == device.id) {
-				return true;
-			}
-		}
-	}
-	return false;
-}
-
-function getFriendlyName(serviceTxt) {
-	if (!serviceTxt) {
-		log('debug-error', 'getFriendlyName()', 'service.txt is missing');
-		return;
-	}
-	var fns = serviceTxt.filter(function (txt) {
-		return txt.match(/fn=*/)!=null;
-	});
-	if (fns.length>0) {
-		var fn=fns[0];
-		log('debug', 'getFriendlyName()', 'is friendly name: ' + fn);
-		return (fn.replace(/fn=*/, ''));
-	} else {
-		log('debug', 'getFriendlyName()', 'is not friendly name: ' + fn);
-	}
-}
-
-function getId(id) {
-	if (id&&id!=null&&id.match(/id=*/)!=null) {
-		log('debug', 'getId()', 'is id: ' + id);
-		return (id.replace(/id=*/, ''));
-	} else {
-		log('debug', 'getId()', 'is not id: ' + id);
-	}
-}
-
-function getGroupIds(serviceTxt) {
-	var groupIds = [];
-	serviceTxt.forEach(function(element) {
-		try {
-			if (!element.includes('id') && !element.includes('__common_time__')) {
-				var groupId = element.split('|', 1)[0].split('=', 1)[0]
-				log( 'debug', 'getGroupIds()', 'memberId: ' +  groupId);
-				if (groupId && groupId != '') {
-					groupIds.push(groupId);
-				}
-			}
-		} catch (e) {
-			log( 'error', 'getGroupIds()', 'cannot get member, error: ' + e );
-		}
-	});
-	return groupIds;
-}
-
-function getNewRequestId(){
-	if(currentRequestId > 9998){
-		currentRequestId=1;
-		log('debug', 'getNewRequestId()', 'reset');
-	}
-	log('debug', 'getNewRequestId()', currentRequestId+1);
-	return currentRequestId++;
 }
 
 function getLatestVersion() {
